+++ src/tattler/client/tattler_py/serialization.py
@@ -1,4 +1,8 @@
+"""Serialization logic to transport client objects faithfully over JSON to server"""
+
 from datetime import datetime, date
+
+from collections.abc import Mapping, Set
 
 import json
 
@@ -7,7 +11,9 @@ class JSONComplEncoder(json.JSONEncoder):
     def default(self, obj):
         if isinstance(obj, datetime) or isinstance(obj, date):
             return obj.isoformat()
-        if isinstance(obj, set):
+        if isinstance(obj, Mapping):    # e.g. SQLAlchemy Row._mapping
+            return dict(obj)
+        if isinstance(obj, Set):
             return sorted(obj)
 
 def serialize_json(dictionary):