+++ src/tattler/server/tests/fixtures/plugins_raising/good_tattler_plugin.py
@@ -1,0 +1,5 @@
+from tattler.server.pluginloader import ContextPlugin, ContextType
+
+class GoodTattlerPlugin(ContextPlugin):
+    def process(self, context: ContextType) -> ContextType:
+        return context | {'foo': 123}