+++ src/tattler/server/tests/test_pluginloader.py
@@ -62,22 +62,23 @@ class PluginLoaderTest(unittest.TestCase):
 
     def test_plugins_initialized_in_alphanumeric_order(self):
         """Plug-in classes are initialized in alphanumeric order of their class name"""
-        path = str(Path(__file__).parent / 'fixtures/plugin_loading_order')
-        loaded_plugins = pluginloader.load_plugins_from_modules([path])
+        fpath = Path(__file__).parent / 'fixtures/plugin_loading_order'
+        loaded_plugins = pluginloader.load_plugins_from_modules([str(fpath)])
         self.assertIn('context', loaded_plugins)
         self.assertEqual(['OneTattlerPlugin', 'ThreeTattlerPlugin', 'TwoTattlerPlugin'], list(loaded_plugins['context'].keys()))
 
     def test_only_valid_plugins_are_initialized(self):
-        """Only plugins respecting their interface are loaded, and have their setup() function called"""
+        """Each valid plugin has its setup() function called during loading"""
+
         with mock.patch('tattler.server.pluginloader.load_candidate_modules') as mcandmod:
             with mock.patch('tattler.server.pluginloader.importlib'):
                 with mock.patch('tattler.server.pluginloader.plugin_category') as misplug:
-                    with mock.patch('tattler.server.pluginloader.inspect') as minsp:
+                    with mock.patch('tattler.server.pluginloader.inspect.getmembers') as mgetmem:
                         with mock.patch('tattler.server.pluginloader.check_sanity'):
                             mcandmod.return_value = {f'module{i}{pluginloader.plugins_suffix}':None for i in range(1, 5)}
-                            minsp.getmembers.side_effect = [
+                            mgetmem.side_effect = [
                                 [[f'classname{i}', mock.MagicMock()]]
-                                for i, n in enumerate(mcandmod.return_value)] # return different values at each subsequent call
+                                for i, _ in enumerate(mcandmod.return_value)] # return different values at each subsequent call
                             misplug.side_effect = lambda n, cl: 'context' if ('1' in n or '3' in n) else None
                             plugcand = pluginloader.load_candidate_modules(self.plugin_paths)
                             plugmap = pluginloader.load_plugins(plugcand)
@@ -87,29 +88,16 @@ class PluginLoaderTest(unittest.TestCase):
                                 pobj.setup.assert_called()
 
     def test_plugins_failing_sanity_check_are_not_loaded(self):
-        """Plug-in classes that fail sanity_check() are not loaded"""
-        with mock.patch('tattler.server.pluginloader.load_candidate_modules') as mcandmod:
-            with mock.patch('tattler.server.pluginloader.importlib'):
-                with mock.patch('tattler.server.pluginloader.plugin_category') as misplug:
-                    with mock.patch('tattler.server.pluginloader.inspect') as minsp:
-                        with mock.patch('tattler.server.pluginloader.check_sanity') as mcs:
-                            modnames = {f'module{i}{pluginloader.plugins_suffix}':None for i in range(1, 5)}
-                            mcandmod.return_value = modnames
-                            mcs.side_effect = [True for _ in list(modnames)[:-1]] + [False]
-                            minsp.getmembers.side_effect = [
-                                [[f'classname{i}', mock.MagicMock()]]
-                                for i, n in enumerate(mcandmod.return_value)] # return different values at each subsequent call
-                            misplug.side_effect = lambda n, cl: 'context' if ('1' in n or '3' in n) else None
-                            plugcand = pluginloader.load_candidate_modules(self.plugin_paths)
-                            plugmap = pluginloader.load_plugins(plugcand)
-                            self.assertEqual({'context'}, set(plugmap.keys()))
-                            self.assertEqual(2, len(plugmap['context']))
-                            for _, pobj in plugmap['context'].items():
-                                pobj.setup.assert_called()
+        """Plug-in classes that fail check_sanity() are not loaded"""
+        plugpath = Path(__file__).parent / 'fixtures' / 'plugins_with_bad_inheritance'
+        plugcand = pluginloader.load_candidate_modules([str(plugpath)])
+        plugmap = pluginloader.load_plugins(plugcand)
+        # if there is a context, there should be no 'GoodTattlerPlugin' in it
+        self.assertNotIn('context', plugmap)
 
     def test_deprecated_contextclassname_loaded(self):
         """Plug-ins using deprecated class name are still loaded."""
-        fpath = Path(__file__).parent / 'fixtures/plugins_sanity_check'
+        fpath = Path(__file__).parent / 'fixtures' / 'plugins_sanity_check'
         loaded_plugins = pluginloader.load_plugins_from_modules([str(fpath)])
         self.assertIn('context', loaded_plugins)
         self.assertIn('WarnTattlerPlugin', loaded_plugins['context'])
@@ -131,18 +119,20 @@ class PluginLoaderTest(unittest.TestCase):
             self.assertIn('ContextTattlerPlugin', mlogwarn.call_args.args[0])
 
     def test_init_tolerates_failing_plugins(self):
-        with mock.patch('tattler.server.pluginloader.plugin_category') as misplug:
-            with mock.patch('tattler.server.pluginloader.inspect') as minsp:
-                with mock.patch('tattler.server.pluginloader.check_sanity') as mcs:
-                    pname = f'module1{pluginloader.plugins_suffix}'
-                    mcandmod = {pname:None}
-                    mockpluginclass = mock.MagicMock()
-                    mockpluginclass().setup.side_effect = RuntimeError
-                    misplug.return_value = 'context'
-                    minsp.getmembers.return_value = [['foo', mockpluginclass]]
-                    plugmap = pluginloader.load_plugins(mcandmod)
-                    mockpluginclass().setup.assert_called()
-                    self.assertEqual(0, len(plugmap))
+        """init() does not fail if a plug-in raises exception while loading"""
+        fpath = Path(__file__).parent / 'fixtures' / 'plugins_raising'
+        with mock.patch('tattler.server.pluginloader.log') as mlog:
+            loaded_plugins = pluginloader.load_plugins_from_modules([str(fpath)])
+            self.assertIn('context', loaded_plugins)
+            self.assertIn('addressbook', loaded_plugins)
+            self.assertIn('GoodTattlerPlugin', loaded_plugins['context'])
+            self.assertIn('WorkingContextPlugin', loaded_plugins['context'])
+            self.assertIn('WorkingAddressbookPlugin', loaded_plugins['addressbook'])
+            self.assertNotIn('RaisesInitializationContextPlugin', loaded_plugins['context'])
+            self.assertNotIn('RaisesInitializationAddressbookPlugin', loaded_plugins['addressbook'])
+            mlog.error.assert_called()
+            setupfailmsg = [call for call in mlog.error.call_args_list if 'failed to setup' in call.args[0]]
+            self.assertEqual({'RaisesInitializationContextPlugin', 'RaisesInitializationAddressbookPlugin'}, set(c.args[1] for c in setupfailmsg))
 
     def test_process_context_runs_all_enabled_plugins(self):
         """All enabled context plugins are executed when processing"""